update transaction status is merged with order service
