+++ orchestrator/config.py
@@ -24,7 +24,7 @@ SLACK_HOOK = {}
 CONTRACT_API_ARN = ""
 ORDER_DETAILS_ORDER_ID_ARN = ""
 ORDER_DETAILS_BY_USERNAME_ARN = ""
-REGION_NAME = ""
+REGION_NAME = "us-east-1"
 SIGNER_ADDRESS = ""
 EXECUTOR_ADDRESS = ""
 SIGNER_SERVICE_ARN = ""
